# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AABRDP blockchain",
    "description": "AABRDP blockchain client",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/Bojidar00/polygon-explorer-api/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://localhost:1337",
      "description": "Local server"
    }
  ],
  "paths": {
    "/chain": {
      "get": {
        "tags": [
          "GET data"
        ],
        "summary": "Get blockchain info",
        "description": "Get blockchain info",
        "responses": {
          "200": {
            "description": "Fetched blockchain info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Fail fetching blockchain"
          }
        }
      }
    },
    "/generate": {
      "get": {
        "tags": [
          "Key pair"
        ],
        "summary": "Generate key pair",
        "description": "Generate key pair with ECDSA - secp256k1. Files are stored in key pairs directory.",
        "responses": {
          "200": {
            "description": "Private and public key data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/blocks/latest": {
      "get": {
        "tags": [
          "GET data"
        ],
        "summary": "Get last block",
        "description": "Get last block from blockchain",
        "responses": {
          "200": {
            "description": "Get latest block object parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": [
          "GET data"
        ],
        "summary": "Get pending transactions",
        "description": "Get pending transactions from blockchain",
        "responses": {
          "200": {
            "description": "Get pending transactions array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/address/balance/{address}": {
      "get": {
        "tags": [
          "GET data"
        ],
        "summary": "Get balance of address",
        "description": "Get balance of address in blockchain",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "string public key",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "04f0d12eaf2c20da771ccfc8c86f3d26e445a44b0c1d8574d2232a3d2c50bc3933fdf9af2e0c36e444d1b53afbcbd6488203336b235b1b445f888c8abd5dbaae8f"
          }
        ],
        "responses": {
          "200": {
            "description": "Get balance of address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/addtransaction": {
      "post": {
        "tags": [
          "Add transaction"
        ],
        "summary": "Add transaction",
        "description": "Add transaction to the blockchain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addtransaction_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Show current pending transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/minetransactions": {
      "post": {
        "tags": [
          "Mine transactions"
        ],
        "summary": "Mine transactions",
        "description": "Mine pending transactions in the blockchain. New block with transactions will be added.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/minetransactions_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Show current chain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          }
        }
      }
    },
    "/getcoins": {
      "post": {
        "tags": [
          "Get coins"
        ],
        "summary": "Get coins",
        "description": "Get coins to address for testing purposes.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getcoins_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Show informative message.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "addtransaction_body": {
        "type": "object",
        "properties": {
          "fromAddress": {
            "type": "string",
            "example": "04f0d12eaf2c20da771ccfc8c86f3d26e445a44b0c1d8574d2232a3d2c50bc3933fdf9af2e0c36e444d1b53afbcbd6488203336b235b1b445f888c8abd5dbaae8f"
          },
          "toAddress": {
            "type": "string",
            "example": "testAddress"
          },
          "amount": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "minetransactions_body": {
        "type": "object",
        "properties": {
          "mineraddress": {
            "type": "string",
            "example": "04f0d12eaf2c20da771ccfc8c86f3d26e445a44b0c1d8574d2232a3d2c50bc3933fdf9af2e0c36e444d1b53afbcbd6488203336b235b1b445f888c8abd5dbaae8f"
          }
        }
      },
      "getcoins_body": {
        "type": "object",
        "properties": {
          "toAddress": {
            "type": "string",
            "example": "04f0d12eaf2c20da771ccfc8c86f3d26e445a44b0c1d8574d2232a3d2c50bc3933fdf9af2e0c36e444d1b53afbcbd6488203336b235b1b445f888c8abd5dbaae8f"
          }
        }
      },
      "inline_response_200": {
        "oneOf": [
          {
            "type": "object"
          },
          {
            "type": "string"
          }
        ]
      }
    }
  }
}